pyproject.toml
[project]
name = "housing-market-simulation"
version = "0.1.0"
description = "Housing Market Simulation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=37.3.0",
    "frictionless[pandas]>=5.18.0",
    "jupyter-book<2",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "pandera>=0.30.1",
    "pyarrow>=19.0.0",
    "pydantic-settings>=2.13.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
    "tqdm>=4.66.0",
]

[dependency-groups]
dev = [
    "wheel>=0.46.3",
]
lint = [
    "pre-commit>=4.5.1",
]


#########################
# Tasks
#########################

[tool.poe.tasks]
serve = "uv run python -m upskills"
format = "uv run pre-commit run --all-files --verbose"
lint = "uv run ruff check src"
typecheck = "uv run mypy src"
docker-build = "docker buildx build --progress plain -t upskills-backend ."
docker-run = "docker container run --env-file .env --publish 8000:8000 upskills-backend"


#########################
# MyPy
#########################

[tool.mypy]
strict = true
follow_imports = "normal"
python_version = "3.12"
disallow_any_unimported = true
disallow_any_generics = true
disallow_subclassing_any = true
ignore_missing_imports = false
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_untyped_decorators = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "passlib.*",
    "jose.*",
    "dependency_injector.*",
]
ignore_missing_imports = true


#########################
# RUFF
#########################

[tool.ruff]
fix = true
preview = true
unsafe-fixes = true
line-length = 120

[tool.ruff.lint]
extend-select = ["ALL"]
ignore = [
    "D",      # pydocstyle
    "CPY",    # copyright notices
    "S101",   # assert usage
    "ISC001", # implicit string concat
    "ANN401", # any type
    "COM812", # trailing comma
    "E501",   # line too long (handled by formatter)
    "B008",   # function calls in defaults (FastAPI Depends)
    "ARG002", # unused method argument
    "RUF067", # __init__ module logic (FastAPI router pattern)
    "S105",   # hardcoded password (dev defaults in pydantic settings)
    "S107",   # hardcoded password in default (false positive on token_type)
    "DOC501", # raised exception missing from docstring
    "DOC502", # documented exception from inner function (API docs clarity)
]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"


#########################
# Pylint
#########################

[tool.pylint.main]
extension-pkg-allow-list = ["mypy","dependency_injector","pydantic_settings"]

[tool.pylint.typecheck]
generated-members = ["sqlalchemy.func.*"]
contextmanager-decorators = ["contextlib.contextmanager"]
ignore-none = true
ignore-on-opaque-inference = true
ignored-checks-for-mixins = [
    "no-member",
    "not-async-context-manager",
    "not-context-manager",
    "attribute-defined-outside-init",
]
ignored-classes = ["optparse.Values", "thread._local", "_thread._local"]
missing-member-hint = true
missing-member-hint-distance = 1
missing-member-max-choices = 1
mixin-class-rgx = ".*[Mm]ixin"

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "C0301",  # line-too-long (handled by ruff format)
    "E0401",  # import-error (already validated by mypy)
    "R0801",  # duplicate-code (acceptable for response builders)
    "R0903",  # too-few-public-methods (dataclasses/DTOs)
    "R0913",  # too-many-arguments (DI constructors)
    "R0917",  # too-many-positional-arguments (DI constructors)
]


#########################
# Bandit
#########################

[tool.bandit]
skips = ["B101"]
